requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
blas=1.0=openblas
blinker=1.6.2=py312hca03da5_0
bottleneck=1.3.7=py312ha86b861_0
bzip2=1.0.8=h80987f9_6
ca-certificates=2024.9.24=hca03da5_0
click=8.1.7=py312hca03da5_0
expat=2.6.3=h313beb8_0
flask=3.0.3=py312hca03da5_0
flask-cors=5.0.0=pypi_0
itsdangerous=2.2.0=py312hca03da5_0
jinja2=3.1.4=py312hca03da5_0
kafka-python-ng=2.2.3=pypi_0
libcxx=14.0.6=h848a8c0_0
libffi=3.4.4=hca03da5_1
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libopenblas=0.3.21=h269037a_0
llvm-openmp=14.0.6=hc6e5704_0
markupsafe=2.1.3=py312h80987f9_0
ncurses=6.4=h313beb8_0
numexpr=2.8.7=py312h0f3ea24_0
numpy=1.26.4=py312h7f4fdc5_0
numpy-base=1.26.4=py312he047099_0
openssl=3.0.15=h80987f9_0
pandas=2.2.2=py312hd77ebd4_0
pip=24.2=py312hca03da5_0
python=3.12.7=h99e199e_0
python-dateutil=2.9.0post0=py312hca03da5_2
python-tzdata=2023.3=pyhd3eb1b0_0
pytz=2024.1=py312hca03da5_0
readline=8.2=h1a28f6b_0
setuptools=75.1.0=py312hca03da5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.45.3=h80987f9_0
tk=8.6.14=h6ba3021_0
tzdata=2024b=h04d1e81_0
werkzeug=3.0.3=py312hca03da5_0
wheel=0.44.0=py312hca03da5_0
xz=5.4.6=h80987f9_1
zlib=1.2.13=h18a0788_1
